+++ src/ui/cards/model/index.js
@@ -42,3 +42,7 @@ export {
 } from './digishelf.js';
 export { default as buildShopilyModel, selectNiche as selectShopilyNiche } from './shopily.js';
 export { default as buildTrendsModel } from './trends.js';
+export {
+  default as buildServerHubModel,
+  selectServerHubNiche
+} from './serverhub.js';